+++ EliteJournalReader/Events/ModuleBuyEvent.cs
@@ -9,13 +9,13 @@ namespace EliteJournalReader.Events
 {
     //When Written: when buying a module in outfitting
     //Parameters:
-    //•	Slot: the outfitting slot
-    //•	BuyItem: the module being purchased
-    //•	BuyPrice: price paid
-    //•	Ship: the players ship
+    //ï¿½	Slot: the outfitting slot
+    //ï¿½	BuyItem: the module being purchased
+    //ï¿½	BuyPrice: price paid
+    //ï¿½	Ship: the players ship
     //If replacing an existing module:
-    //•	SellItem: item being sold
-    //•	SellPrice: sale price
+    //ï¿½	SellItem: item being sold
+    //ï¿½	SellPrice: sale price
     public class ModuleBuyEvent : JournalEvent<ModuleBuyEvent.ModuleBuyEventArgs>
     {
         public ModuleBuyEvent() : base("ModuleBuy") { }
@@ -28,7 +28,7 @@ namespace EliteJournalReader.Events
             public string BuyItem_Localised { get; set; }
             public int BuyPrice { get; set; }
             public string Ship { get; set; }
-            public int ShipID { get; set; }
+            public long ShipID { get; set; }
             public string SellItem { get; set; }
             public string SellItem_Localised { get; set; }
             public int? SellPrice { get; set; }
