+++ EliteJournalReader/Events/PayFinesEvent.cs
@@ -9,11 +9,11 @@ namespace EliteJournalReader.Events
 {
     //When Written: when paying fines
     //Parameters:
-    //•	Amount
-    //•	BrokerPercentage (present if paid via a Broker)
-    //•	AllFines: bool
-    //•	Faction: (if paying off an individual faction's fines)
-    //•	ShipID
+    //ï¿½	Amount
+    //ï¿½	BrokerPercentage (present if paid via a Broker)
+    //ï¿½	AllFines: bool
+    //ï¿½	Faction: (if paying off an individual faction's fines)
+    //ï¿½	ShipID
 
     public class PayFinesEvent : JournalEvent<PayFinesEvent.PayFinesEventArgs>
     {
@@ -26,7 +26,7 @@ namespace EliteJournalReader.Events
             public bool AllFines { get; set; }
             public string Faction { get; set; }
             public string Faction_Localised { get; set; }
-            public int ShipID { get; set; }
+            public long ShipID { get; set; }
 
         }
     }
